requirements.txt
Django>=4.2,<6.0
django-filter>=23.0
Pillow>=10.0.0
wikipedia>=1.4.0
requests>=2.31.0

